WIP: so, adversarial validation gets excellent results, but actual test results are very poor, like 0.60 accuracy.
If we treat the test sample as an alternative titanic then I can incorporate known titanic lived, died counts, esp since they hold for both home regions and for ticket amounts. So that shouldn't be the issue but I can test for that.

The only other thing I can think is that the rows have been shuffled, although the PassengerIds still match up to the rest of the row content so I'm not sure either. Defintely strange, need to buckle down here...